+++ src/mugen/characterhud.cpp
@@ -430,7 +430,7 @@ void Bar::render(Element::Layer layer, const Bitmap & bmp){
         middle.render(layer, position.x, position.y, bmp, (int)(damage * range.y / maxHealth));
 
         double width = currentHealth * range.y / maxHealth;
-        front.render(layer, position.x, position.y, bmp, width);
+        front.render(layer, position.x, position.y, bmp, (int) width);
 
 #if 0
         // Middle is the damage indicator
